Minor changes
The cue sections have been added. Issues with the response log have
been identified. For example, the while loop is updating the ‘response’
variable to a ‘miss’ to the point that no other response can be logged.
The reason why pressing ‘space’ works is that it redirects the loop
completely, and it breaks it.

This means that the whole response system within the while loop for
each task function needs to be updated according to the psychopy
tutorial (which didn’t work initially anyways).

The task looks better, but it’s not there yet.