Log semi fixed
Every mental effort condition now goes through the response loop fine,
and sends the final response key to the main loop. If there are more
than 2 incorrect or misses, the loop break.

Problem is that the ‘while’ loop within each condition updates the keys
too fast. Need to set it up so it waits for 1 key press and/or 1.5 secs.